+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
-[unreleased]: https://github.com/rogdham/bigxml/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/rogdham/bigxml/compare/v0.3.0...HEAD
+
+## [0.3.0] - 2020-09-27
+
+[0.3.0]: https://github.com/rogdham/bigxml/compare/v0.2.0...v0.3.0
+
+v0.3.0 improves namespace support.
 
 ### :boom: Breaking changes